I did everything
When a user logs in
- Store their current location as 2dsphere
- query potential users within 11 miles and users that aren't in current
users likes and dislikes array

Created a load page to display to the user while querying for potential
users.

Calculate users age and last time they logged in

Incorporate local storage for userID

Each time a user likes/dislikes another user the userID is added to
their array.