Users can change their bio & Display no potential horses within desired distance
Added a text area on the edit user page to allow users to create there
own bio, limited to 150 characters. Moved the query for potential horses
to the loadCtrl so if there are no potential horses within the desired
constraints a display will be shown to the user telling them so and to
extend the distance of potential matches. The query is on a timeout
remaking the call every 7 seconds to check to see if there is a
potential match. Any time the user leaves the loading page while the
timeout is running a promise is sent to cancel the timeout.